new structure, add a lobbyListComponent, moved css to own file
