+++ app/src/App.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { BrowserRouter as Router, Switch, Route } from 'react-router-dom';
-import Lobby from './components/lobby';
+import LobbyComponent from './components/lobbyComponent/lobbyComponent';
 import Login from './components/login';
 import Game from './components/game';
 
@@ -8,7 +8,7 @@ const App = () => {
   return (
     <Router>
       <Switch>
-        <Route path="/lobby" component={Lobby} />
+        <Route path="/lobby" component={LobbyComponent} />
         <Route path="/game" component={Game} />
         <Route path="/game/:id" component={Game} />
         <Route exact path="/" component={Login} />
