+++ app/src/components/lobbyComponent/lobbyComponent.js
@@ -1,25 +1,15 @@
-import React, { useState, useEffect } from 'react';
+import React, { useState } from 'react';
 import axios from 'axios';
 import { Redirect } from 'react-router-dom';
-import '../App.css';
+import LobbyListComponent from './lobbyListComponent';
+import './lobbyComponent.css';
 
-const Lobby = (props) => {
-  const [games, setGames] = useState([]);
+const LobbyComponent = (props) => {
   const [home, setHome] = useState(false);
   const [playMatch, setPlayMatch] = useState(false);
   const [matchId, setMatchId] = useState('');
   const [errorMess, setErrorMess] = useState('');
   const name = props.location.state.name;
-
-  useEffect(() => {
-    axios.get('/api/seeks', { headers: { 'Content-Type': 'application/json' } })
-      .then((response) => {
-        setGames(response.data);
-      })
-      .catch((error) => {
-        setErrorMess(error);
-      });
-  }, []);
 
   const startGame = () => {
     const player = {
@@ -28,20 +18,6 @@ const Lobby = (props) => {
     axios.post('/api/seeks', JSON.stringify(player), { headers: { 'Content-Type': 'application/json' } })
       .then((response) => {
         setMatchId(response.data.id);
-        setPlayMatch(true);
-      })
-      .catch((error) => {
-        setErrorMess(error);
-      });
-  };
-
-  const joinGame = (id) => {
-    const joinMatch = {
-      spelare: name,
-    };
-    axios.post(`/api/seeks/${id}`, JSON.stringify(joinMatch), { headers: { 'Content-Type': 'application/json' } })
-      .then((response) => {
-        setMatchId(response.data[0].id);
         setPlayMatch(true);
       })
       .catch((error) => {
@@ -78,27 +54,19 @@ const Lobby = (props) => {
               <td>Join Game</td>
             </tr>
           </thead>
-          <tbody>
-            {games.map((game) => {
-              if (game.spelare.length > 1) {
-                return null;
-              }
-              return (
-                <tr key={game.id} className="lobby-tr">
-                  <td>{game.spelare}</td>
-                  <td><button type="button" onClick={() => joinGame(game.id)}>Play</button></td>
-                </tr>
-              );
-            })}
-          </tbody>
+          <LobbyListComponent 
+            name={name} 
+            setMatchId={setMatchId} 
+            setPlayMatch={setPlayMatch} 
+            setErrorMess={setErrorMess} />
         </table>
         <br />
         <button type="button" className="lobby-button" onClick={startGame}>Start New Game</button>
         <button type="button" onClick={logOut}>Log out</button>
-        <div style={{ color: 'red', marginTop: '15px', fontSize: '13px' }}>{errorMess}</div>
+        <div className="lobby-errorMessage">{errorMess}</div>
       </div>
     </div>
   );
 };
 
-export default Lobby;
+export default LobbyComponent;
